requirements/base_requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic-settings==2.5.2

# Database + ORM
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
alembic==1.13.3

# Authentication
pwdlib[argon2]==0.2.0
argon2-cffi>=23.1.0
python-jose[cryptography]==3.3.0

# Validation
email-validator==2.2.0
python-multipart==0.0.12

# HTTP client (for Groq AI API)
httpx>=0.27.0

# Image processing (thumbnails)
Pillow>=10.0.0

# Async file I/O
aiofiles>=23.0.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx>=0.27.0
